+++ CHANGELOG.md
@@ -6,6 +6,7 @@ See [changes since release][HEAD].
 
 * Add "aria-live" to tablists for reporting tab changes to the user.
 * Add optional "aria-label" to tablists to improve screenreader reporting.
+* Give active tabpanels tabindex 0 to allow keyboard navigation to the active panel content.
 
 ## [3.2.0][] / 2018-03-07